services/user/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-environ==0.11.2
psycopg2-binary==2.9.9
celery==5.3.4
redis==5.0.1
django-allauth==0.57.0
django-filter==23.5
drf-spectacular==0.26.5
drf-spectacular-sidecar==2023.10.1
Pillow==10.1.0
python-decouple==3.8
gunicorn==21.2.0
whitenoise==6.6.0
django-redis==5.4.0
